% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Undirbua_kennslu.R
\name{finna_kennslutimabil}
\alias{finna_kennslutimabil}
\title{Listar kennsludaga}
\usage{
finna_kennslutimabil(upphafs_dagur, vikudagar_nr = c(1, 3), fj_vikur = 12)
}
\arguments{
\item{upphafs_dagur}{Dagsettning sem kennsla hefs}

\item{vikudagar_nr}{Vikudagar sem kennt er t.d. ef kennt er a man og mid
vikudagar_nr = c(1, 3)}

\item{fj_vikur}{Fjoldi vikna i namskeidi}
}
\value{
Lista med kennslu dogum
}
\description{
Listar kennsludaga
}
\examples{
finna_kennslutimabil(upphafs_dagur = "2023-12-01", vikudagar_nr = 1, fj_vikur = 3)
}
