% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Undirbua_kennslu.R
\name{formata_daga}
\alias{formata_daga}
\title{Birtir vikudag og manud}
\usage{
formata_daga(dagar)
}
\arguments{
\item{dagar}{dagsettning "2023-12-01"}
}
\value{
Birtir vikudag og manud
}
\description{
Birtir vikudag og manud
}
\examples{
dagar <- c("2023-12-01", "2023-12-08", "2023-12-15")
formata_daga(dagar)
}
