% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smafoll.R
\name{marglida}
\alias{marglida}
\title{Utbua regression jofnu}
\usage{
marglida(p = 0, upper = F)
}
\arguments{
\item{p}{fjoldi parametra X}

\item{upper}{x eda X, defult er upper = FALSE}
}
\description{
Utbua regression jofnu
}
\examples{
marglida(2, upper = T)
marglida()
}
